Complex numbers support, closes #36 (#52)
* fill dataset with NaNs

* specify PySDM version range

* move storage init to external function

* add support for halo > 1

* precommit

* support for complex numbers. closes #36

* precommit

* Rename project (#51)

* specify PySDM version range

* add support for halo > 1

* rename project

* update README

* fix pdoc

---------


* increase timeout

---------

+++ PyMPDATA_MPI/periodic.py
@@ -50,7 +50,8 @@ def _make_send_recv(set_value, jit_flags, fill_buf):
             (
                 len(i_rng),
                 len(k_rng),
-            )
+            ),
+            dtype=output.dtype,
         )
 
         rank = mpi.rank()
